Fixed black box issue and first commit for Diary
I was having an issue with the black box in the top right, but I realised it was caused by defining set heights, rather then using height_hints, that I have now adjusted it to use. I have also removed the DPI scaling since I don't want to rely on that anymore, I want to have the whole window and layhouts scale depending on what size the window is.