Fixed Back-Testing Title
I have fixed the issue I had with the backtesting title, since I now know that I had to define it within the python code since it was being drawn over if I tried to define it in the kv file. Now it is always in the right place and defined properly. Next, I will move on to adding the tabs