Adjusted main layouts to be screens
I realised that I needed to use the "Screen" property to be able to create the tabbing system I want for my program. This will also involve a screenmanager, which I have set up and working, but no way to swap between the screens, and nothing on my other screen. I also needed to define a nested boxlayout within my kivy file, since that was the only way to set up the correct orientation when using a Screen.