First Push of Portfolio Screen
I've spent all day working on this, I wanted to make sure it was in a semi-competent state before I pushed it. I thought it was best to keep with my continued OOP approach by splitting this one up into more classes, but I ended up overcomplicating certain aspects that caused me to try and transfer across too much info when I didn't need to. It currently can have stocks added with no verification, stores them within a Json which is native to kivy and can be clicked on to change the right hand side. I will store the json on git for now, but may change my mind on this in the future. I will also delete it when I need to remove the contents of it for now...