Adjusted Portfolio Logic
I've changed the logic for how my portfolio works, rather then using a bunch of callbacks to parse data between the classes, I've gotten it to store stuff to the JSON and then update the json every time. I am having one problem with all the buttons bar the final one not displaying their labels properly, so I'm aiming to fix that now.