Added yfinance integration for Portfolio
I added yfinance, wanted to get current stock value, but discovered the ticker function for the import is broken, even after checking online. I still discovered a decent way to get as up to date stockvValues as I could, so it didn't impact me too much. With this, I was able to save initial values, which provide a contrast to calculate returns later on. At the moment, I can display accurate totals for everything but VaR for the portfolio, before I implement VaR I will make the menu's more fluid to navigate and display certain bits of information