Set up more Background App Refresh
I set it up so that instead of the stock cards refreshing in the background, which isn't actually really that necessary, I get the right hand side information to refresh, depending if the total values or the specific stock values are needed. This is great as it makes the information the portfolio is providing look fluid and current, and I added some 'race conditions' to make sure that one or the other can't be happening at the same time, so they don't write over each other.