Fixed Button, implemented visual and logical convergence
I first used an anchorLayout to fix where my button was being displayed to go back on the page. I also made it invisible and an X so its easy to use and cane let you return to the overall stocks easier. I also researched and realised that I should be checking for when the MonteCarlo simulations start to converge, so I first created a method that would use matplotlib to show me how it looked, which was great to finally be able to perform visualisations (although not within kivy) but it did not help me understand how many simulations I should perform. I realised I needed to set a convergence point and compare between different VaR's I generate. So I created a new sim method for it that does exactly that, only runs a certain amount of sims depending on when the convergence point is met. I have it commented out now, but I will most likely delete it in a further commit