Exit Popup doesn't run Background App Refresh
The popup, which has a button that is not in the right place, now has it so that the background app refresh of the stock and totals list only happens after you have saved a stock, and not when you are easily exiting from the button. I also removed some unused imports.