Adjusted look and create a Stock Name Finder
Adjusted how the stock buttons look, making them white, and a few other adjustments. I want the stock buttons to display the Name and the Ticker of the stocks, but I can't directly access that information from Yahoo Finance since its broken, so I had to use beautifulsoup to access the yahoopage for it and retrieve the information myself, but it seems to work. I will implement it within the saving process for the stocks soon