Updated to display name + fixed user-agent problem
I was having an issue where it wouldn't load certain foreign (non-American) stock exchange html page results, which didn't let me search up the name for adidas. To combat that, I added a "user-agent", since I found that this helps it mimic a real web browser to get the accurate response. After I got this working, I got the names for my current stocks in my holdings, I added the names to the json, then I got them to be displayed, and initially stored when you try and buy a stock from a ticker. It now looks way nicer when displaying your stocks...