Completely Revamped Previous Stock Prices
I realised that I was individually downloading stocks one by one again, which is always slow and inefficient. To combat this, I realised that whenever I do the background app refresh, I'll always have one of my stock totals being used on the right hand side, or all of them. So when I download the stock for that, I can just parse that total into my stock code. This made the code run so much faster as well. It made me do some adjustments, but it seems to be working great.