Delete button invisible
I made it so that when you are not selecting an individual stock, the delete button is invisible. However, I think I want to get it so that when you click the delete button, it creates a pop-up, where you select yes and no, so I will do that next.