Confirm Delete Popup
I created a confirm deletion popup that will open when you try and delete a stock. It will reload the page if a stock is deleted and handles it all on the popup end. I also adjusted how the other popup looked, so both look and work a lot better now. Next I want to add another popup for when you want to adjust VaR values...