Added saveStock verification
I added verification for saveStock, which was complicated as I had to take into account all different possibilities, for empty strings, one input being correct whilst the other one isn't, etc. It does so, and to make it easier for the user to understand this I added an animation for it to slightly turn red. I think I want to implement this slight animation background colour change whenever I change from initial to specific stock, so I'll keep that in mind