Moved Holdings and imroved VaR Adjustment
I moved my holdings json to the actual folder that I have my program in, I realised that I was running my program from my main project folder which was wrong, so I will have to change that in the future. I also changed what is displayed in my VaR adjust popup so its easier for the user to understand what VaR is, and understand what they need to enter for the popup etc, it looks and works great, now I need to add verification.