Added Var Adjustement Verification
I have now added the verification logic for VarAdjustment. Now it works like the other pop-up, and in the hint text after a failed input, it will tell them the restraints they need to enter for the next one. I have a few things left to do, then I think this screen is done.