Changed Stock Layout, changed background colours, fixed layouts
This should be the last bit I need to do for Portoflio screen for now. I've gotten it in a great position, where the colours aren't too dreadful, you're able to navigate around it without to much of a problem. I had to change the stocks to be a grid layout, so I could get the name and the price to line up together in a satisfying way. I changed the colour and other backgrounds. I adjusted how the right hand side looked and added a small animation when changing between layouts, so its even more obvious.