Final Adjustments Before Trends Tab
I made a few adjustments, like lining up the stocks and holdings to make it easier to understand. I adjusted the left vs right sizing to make it so most quantities can fit in, which is nice. I also cut out VaR decimal places, since for a statistic like it, they're not particularly as relevant