Got Kivy Garden working + made first layout design
I was having a big problem with getting the kivy garden matplotlib import to work, no matter how I tried to link it, the workspace could not see it. So, after searching online, I found some people talking about it, I used a pip install and then changed how I referenced the code, and I got it to work. So thats great, I now have it 100% working for me. So I set up my layout with the scrollable buttons like before, and I tested a fairly basic graph, it all seems to be working initially, idk if I'll be doing more on this today, we shall see.