Buttons now Change Graphs
I got it so that when you click on buttons at the bottom of the Graphs screen, it will change the displayed graph. I was originally doing this without clearing the widgets, so a bunch of graphs squished themselves together, which was funny but I fixed it.  I also fixed some bugs in portfolio.