Changed Display Logic
I changed the display logic so that the code for creating a graph is initialised differently, so that I can put the logic when I call the function, and it can display the results depending on what I give it. So essentially, I can call graph 1 and it will create the numbers for it, pass it in (a factory) and it will deliver the result. This is essentially a factory design pattern, I think, and it seems to work a lot better. With this, I will try and bridge the gap between my two screens with the MonteCarlo Sim.