Updated Kivy and added threading
I updated Kivy being used to the latest version 2.3.0 and I added threading to the app since it would lag badly when I would do the simulations. I have no visual way to show that its loading yet, but I am planning to add one, since it takes a long time now, but using the threading, it lags the program a looot less, so I think it can be worth it if I handle it properly.