Made portfolio a property in graphs
I have made portfolio a property in graphs, so that it I can refer to it in other functions without having to define it every time. I am tempted to do this for holdings in my portfolio class, but since it gets defined in multiple different classes within that same file, I don't think it matters too much, but I want to try and use this again in the future.