Added Past Theoretical Stock Value & Custom Scroll Wheel
I added the past theoretical stock value by going back and finding all the closing values for the stocks downloaded and passed in, this will then all be added together for each day of the last 500 days, then plotted every 15 days on the graph. There seem to be a bunch of gaps, so I need to fix that, since it looks terrible displayed. I also created a custom scroll wheel, since the one I was using was terrible, this new scrollwheel looks much better.