Fixed missing values and added a symbol
I used if statements to check for missing value, and creating estimated ones based as such, I think they're logical assumptions and make the graphs look better. I also added a symbol that is displayed, although I think I need to make sure the logic is more robust for it. I also added borders to some graphs in the scroll section but I think I need to edit it more.