Threading Stopper, Inverting Axis & Vectorisation Improvements
I was having problems with my Threading, where it wouldn't let me close the program when it was active, so I added a way to let me. I wanted to have the past stock history axis show numbers from 500-0, so I adapted it so it would check and do so, I think the values that it now displays are pretty good. Finally, I was getting annoyed at how long my simulations took, so I adapted the code to use vectorisations and it improved dramatically, to the point where it works very fast in my graphing and portfolio page.