Created a Selected Stock Graph + updated final y points
I created another graph that would show the change in shares price over the last 500 days for it, this is a standard thing to have within many stock programs so I thought I should add it, I have it verifying that a stock has been selected from the previous screen so that it can only run then. I also changed the final y values that are plotted on this graph and the total value graph to be the values displayed on the portfolio screen, to maintain consistency.