Seperated Nan logic, created an Image Blur-er
The nan-logic was large and the same for two of my graphs, so I thought I would make it its own callable method, which seems to have worked, although I have to round all my numbers now haha. I also know that in the future I want to blur out the graphs below, so I can display text in front of them that's easier to read, so I've made a python application to do so, I tested it and I think it can save and work for what I need of it in the future.