FTSE Graph shows Company Names
I decided it would look better for company names, so I took the list, grabbed the names and zipped them together, which allowed me to pass in the correct values when displaying, and now it shows the company names. I need to change it so that they don't get stuck unreadable beyond the graphs boundaries.