Backtesting MonteCarlo Sim and More Bugfixing
I performed back testing for my montecarlosim, it worked, albeit very slow, and didn't increment its count once, so I don't know if I've done something wrong with it either. I then realised that I was having a lot of bugs, first it didn't let me, with my new faster montecarlosim approach, adjust the time horizon anymore. I managed to fix this, it was to do with the dimensions not aligning properly for a numpy sum. I also realised that when I delete all of my stocks, a loooot of things start to break in my code, so I need to go around and fix all of that. I fixed a few errors about when my stocks list has only a single stock in it as well.