Edited Backtesting to look better and general code cleaning
I've got the backtesting so that it now displays the correct x-values. I added the custom scrollbar to the other classes that needed it. I also generally have cleaned up the code. I'm going to finish up on the graphing, clean up the code, then I'll update my diary and start on the report