Changed Ranking axis to % and gave graphs different colours
I made it so that the ranking graphs show the percentage of value at risk, since I think this is a better metric to compare against rather then having to display it with an arbitrary value. I also have added custom colours to all the graphs, so they are distinct.