Comments for VaRChecker and Final Design & Cleaning
I added and adjusted comments for my last term deliverable, VaRChecker, and I cleaned up the code as well in places, such as deleting sections that are now not needed. I also added a few comments to my Final Design main hub as well, although it doesn't need many. I think the final thing I want to do is get the animations to change depending on what direction you are clicking on the tabs in. After that, I will compile my program and see if it works.