Added JSON info and dynamic Screen Sliding
I didn't like the fact that the screen would slide only one direction, it would always have the same animation no matter where the screen was sliding too. To fix this, I had it so it would check what index it was currently in, then which one was being selected, and change the animation depending on this, which looks brilliant. I also added a little bit of information on how the JSON holdings works, since it's useful.