Managed to deploy an Exe
I managed to get a deployed working exe file, but originally it kept crashing and not working at all. I realised that because of the way that it gets packaged within pyinstaller, it would need to have specific pathing introduced to get it to work with my other files that I've defined and put in as folders. So I set this up, having to search up how to define it properly, and got it to register my kv files properly, since I couldn't like them within my kivy file like I had done so previously. I also needed to specify the same path for my png's, so I needed to define a PATH recourse that could be used within other kv files. It is deployable with console, but I want to see if I can do it without. I can't commit the big file to github so I won't for now...