Made a working no-console distribution of the software
I discovered that the reason I could never get the distributions to work without a console is there are imports, such as yfinance, that need somewhere to put there logging infomration, or they break. To combat this, I set the app to immediately check if it is in a development enviroment, and if not, it will start logging to a bin.log file. This logging does accumulate over time, but its very small kb of data, and it overwrites it whenever the app starts up. I think this is the perfect way to have it be distributed as a single exe, although I may try a way to do it that reduces the size of the exe file